# client/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/admin/db":{"delete":{"tags":["Admin"],"summary":"Delete Db","description":"Drop the database.","operationId":"drop_db","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Drop Db"}}}}}}},"/campaigns/":{"get":{"tags":["Campaigns"],"summary":"All Campaigns","description":"Get all campaigns","operationId":"get_all_campaigns","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Campaign"},"type":"array","title":"Response Get All Campaigns"}}}}}},"post":{"tags":["Campaigns"],"summary":"Create Campaign","description":"Create a new campaign","operationId":"create_campaign","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/campaigns/{id}":{"delete":{"tags":["Campaigns"],"summary":"Delete Campaign","description":"Delete a campaign","operationId":"delete_campaign","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Campaigns"],"summary":"Update Campaign","description":"Update a campaign","operationId":"update_campaign","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CampaignUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Campaign"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/{campaign}":{"get":{"tags":["Messages"],"summary":"Get Campaign Messages","description":"Get all messages for a campaign","operationId":"get_campaign_messages","parameters":[{"required":true,"schema":{"type":"string","title":"Campaign","example":"5eb7cf5a86d9755df3a6c593"},"name":"campaign","in":"path"},{"required":false,"schema":{"type":"integer","title":"Skip","default":0},"name":"skip","in":"query"},{"required":false,"schema":{"type":"integer","title":"Limit","default":100},"name":"limit","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Message"},"type":"array","title":"Response Get Campaign Messages"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Messages"],"summary":"Send Message","description":"Create a new message.","operationId":"send_message","parameters":[{"required":true,"schema":{"type":"string","title":"Campaign"},"name":"campaign","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/{id}":{"delete":{"tags":["Messages"],"summary":"Delete Message","description":"Delete a message","operationId":"delete_message","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Messages"],"summary":"Update Message","description":"Update a message.","operationId":"update_message","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Message"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/":{"get":{"tags":["Users"],"summary":"All Users","description":"Get all users","operationId":"get_all_users","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/User"},"type":"array","title":"Response Get All Users"}}}}}},"post":{"tags":["Users"],"summary":"Create User","description":"Create a new user","operationId":"create_user","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/{id}":{"delete":{"tags":["Users"],"summary":"Delete User","description":"Delete a user","operationId":"delete_user","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Users"],"summary":"Update User","description":"Update a user","operationId":"update_user","parameters":[{"required":true,"schema":{"type":"string","title":"Id"},"name":"id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"tags":["Root"],"summary":"Get Root","description":"Root endpoint for the API.","operationId":"get_root","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","title":"Response Get Root"}}}}}}}},"components":{"schemas":{"Campaign":{"properties":{"_id":{"type":"string","title":" Id","example":"5eb7cf5a86d9755df3a6c593"},"name":{"type":"string","title":"Name"}},"type":"object","required":["_id","name"],"title":"Campaign","description":"Output of campaign data"},"CampaignCreate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"CampaignCreate","description":"New campaign data"},"CampaignUpdate":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","title":"CampaignUpdate","description":"Update of campaign data"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Message":{"properties":{"_id":{"type":"string","title":" Id","example":"5eb7cf5a86d9755df3a6c593"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"campaign":{"type":"string","title":"Campaign","example":"5eb7cf5a86d9755df3a6c593"},"sender":{"type":"string","title":"Sender","example":"5eb7cf5a86d9755df3a6c593"},"target":{"type":"string","title":"Target","example":"5eb7cf5a86d9755df3a6c593"},"content":{"type":"string","title":"Content"},"is_private":{"type":"boolean","title":"Is Private"},"is_edited":{"type":"boolean","title":"Is Edited"}},"type":"object","required":["_id","timestamp","campaign","sender","content","is_private","is_edited"],"title":"Message","description":"Output of message data"},"MessageCreate":{"properties":{"campaign":{"type":"string","title":"Campaign","example":"5eb7cf5a86d9755df3a6c593"},"sender":{"type":"string","title":"Sender","example":"5eb7cf5a86d9755df3a6c593"},"target":{"type":"string","title":"Target","example":"5eb7cf5a86d9755df3a6c593"},"content":{"type":"string","title":"Content"},"is_private":{"type":"boolean","title":"Is Private","default":false}},"type":"object","required":["campaign","sender","content"],"title":"MessageCreate","description":"New message data"},"MessageUpdate":{"properties":{"campaign":{"type":"string","title":"Campaign","example":"5eb7cf5a86d9755df3a6c593"},"sender":{"type":"string","title":"Sender","example":"5eb7cf5a86d9755df3a6c593"},"target":{"type":"string","title":"Target","example":"5eb7cf5a86d9755df3a6c593"},"content":{"type":"string","title":"Content"},"is_private":{"type":"boolean","title":"Is Private"}},"type":"object","title":"MessageUpdate","description":"Update of message data"},"User":{"properties":{"_id":{"type":"string","title":" Id","example":"5eb7cf5a86d9755df3a6c593"},"username":{"type":"string","title":"Username"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"avatar":{"type":"string","title":"Avatar"},"is_admin":{"type":"boolean","title":"Is Admin"}},"type":"object","required":["_id","username","name","email","password","is_admin"],"title":"User","description":"Output of user data"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"avatar":{"type":"string","title":"Avatar"}},"type":"object","required":["username","name","email","password"],"title":"UserCreate","description":"New user data"},"UserUpdate":{"properties":{"username":{"type":"string","title":"Username"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"avatar":{"type":"string","title":"Avatar"},"is_admin":{"type":"boolean","title":"Is Admin"}},"type":"object","title":"UserUpdate","description":"Update of user data"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}